Modify simpleWatermark and simpleWatermarkTest based on comments from TimestampWatermarkTest
RB=419259
R=cbotev,mitu,ydai,ynli
A=cbotev

+++ source/src/main/java/com/linkedin/uif/source/extractor/watermark/Watermark.java
@@ -35,7 +35,7 @@ public interface Watermark {
    * @param maximum number of partitions
    * @return partitions
    */
-  public HashMap<Long, Long> getIntervals(long lowWatermarkValue, long highWatermarkValue, int partitionInterval,
+  public HashMap<Long, Long> getIntervals(long lowWatermarkValue, long highWatermarkValue, long partitionInterval,
       int maxIntervals);
 
   /**
